requirements.txt
werkzeug==2.3.7
Flask~=3.0.0
flask_sqlalchemy
flask_login
flask_wtf
WTForms~=3.1.0
flask-qrcode
pyotp~=2.9.0
markupsafe~=2.1.3
cryptography~=41.0.7
rsa~=4.9
email_validator
bcrypt~=4.1.1
SQLAlchemy~=2.0.21
python-dotenv~=1.0.0
flask-talisman